# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OfferMaker API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/offers": {
            "get": {
                "tags": [
                    "Offer"
                ],
                "summary": "Offres",
                "description": "get all offres",
                "operationId": "3a56496f000faffd87f42c4036770f0e",
                "responses": {
                    "200": {
                        "description": "get all offres"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "No offers found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Offer"
                ],
                "summary": "Offres",
                "description": "get all offres",
                "operationId": "4b4cc5dac461b7ed0b8d56a8a2ce22d2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "location",
                                    "company_name",
                                    "salary",
                                    "job_type",
                                    "deadline",
                                    "name"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "location": {
                                        "type": "string"
                                    },
                                    "company_name": {
                                        "type": "string"
                                    },
                                    "salary": {
                                        "type": "number"
                                    },
                                    "job_type": {
                                        "type": "string",
                                        "default": "full-time"
                                    },
                                    "deadline": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "add offres"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}